requirements.txt
# Core dependencies
kafka-python==2.0.2
requests==2.31.0
pydantic==2.5.0
retrying==1.3.4
python-dotenv==1.0.0

# Enhanced monitoring and performance
psutil==5.9.6

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Data processing and validation
jsonschema==4.20.0
pandas==2.1.4
numpy==1.24.4

# Logging and monitoring
prometheus-client==0.19.0

# Security and configuration
cryptography==41.0.8
pyyaml==6.0.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1